ver. 2.5.1: cleaning up names of components
