+++ src/app/spotify/components/topArtistList_functions.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useMemo, useCallback, useEffect } from "react";
-import { TimeRangeRadioButtons } from "@/app/spotify/components/userTopArtist/components/timeRangeRadioButtons_component";
+import { TimeRangeRadioButtons } from "@/app/spotify/components/userTopArtist/components/topArtistTimeRangeButtons_component";
 import { TopArtistList } from "./userTopArtist/components/topArtistListRecords_component";
 import { UserTopArtistsAndTracks } from "@/app/spotify/interfaces/getUserTopArtistsAndTracks";
 import { Pagination } from "./userTopArtist/components/topArtistListPagination_component";
