+++ src/app/spotify/components/userTopArtist/components/userTopArtistList_containerComponent.tsx
@@ -1,31 +1,64 @@
 "use client";
-import React, { useState, useMemo, useCallback, useEffect } from "react";
-// import "./spotify.css";
+import React, { useState, useMemo, useEffect } from "react";
 import "@/app/spotify/components/spotify.css";
 import { useGetTopArtistsFullList } from "@/app/spotify/hooks/spotifyHooks";
-import { TimeRangeRadioButtonsOnChange } from "@/app/spotify/components/userTopArtist/components/timeRangeRadioButtons_component";
+import { TopArtistTimeRangeButtonsOnChange } from "@/app/spotify/components/userTopArtist/components/topArtistTimeRangeButtons_component";
 import { ArtistPaginationOnChange } from "@/app/spotify/components/userTopArtist/components/topArtistListPagination_component";
-import { TopArtistListOnChange } from "@/app/spotify/components/userTopArtist/components/topArtistListRecords_component";
+import { TopArtistListRecordsOnChange } from "@/app/spotify/components/userTopArtist/components/topArtistListRecords_component";
 
-interface UserTopArtistService {
+interface UserTopArtistCardProps {
   pageSize: number;
   defaultTimeRange: string;
   time_ranges: string[];
 }
 
-export function UserTopArtistService({ pageSize, time_ranges, defaultTimeRange }: UserTopArtistService) {
+////////////////////////////
+// Wrapper Component ///////
+////////////////////////////
+export function UserTopArtistCard({ pageSize, time_ranges, defaultTimeRange }: UserTopArtistCardProps) {
+  const { userTopArtistPresentation: userTopArtistCard } = UserTopArtistPresentation({
+    pageSize,
+    time_ranges,
+    defaultTimeRange,
+  });
+  return { userTopArtistCard };
+}
+
+////////////////////////////
+// Presentation Component //
+////////////////////////////
+
+function UserTopArtistPresentation({ pageSize, time_ranges, defaultTimeRange }: UserTopArtistCardProps) {
+  const { timeRangeRadioButtonsComponent, topArtistListOnChangeComponent, paginationComponent } =
+    UserTopArtistContainer({
+      pageSize,
+      time_ranges,
+      defaultTimeRange,
+    });
+  const userTopArtistPresentation = (
+    <div className="top-artist-container">
+      <header className="top-artist-container-header">
+        Top Artists
+        {timeRangeRadioButtonsComponent}
+      </header>
+      <main className="top-artist-list-container">{topArtistListOnChangeComponent}</main>
+      <footer className="top-artist-container-footer">{paginationComponent}</footer>
+    </div>
+  );
+
+  return { userTopArtistPresentation };
+}
+
+////////////////////////////
+// Container Component /////
+////////////////////////////
+function UserTopArtistContainer({ pageSize, time_ranges, defaultTimeRange }: UserTopArtistCardProps) {
   const [resetTrigger, setResetTrigger] = useState(false); // New state to act as a reset trigger...used to let child component that topartistsfulllist has changed, radio button change was too fast and made page number reset to 1 before topartistsfulllist was updated
 
-
-  const { timeRangeRadioButtonsComponent, selectedTimeRangeButton } = TimeRangeRadioButtonsOnChange({
+  const { timeRangeRadioButtonsComponent, selectedTimeRangeButton } = TopArtistTimeRangeButtonsOnChange({
     defaultTimeRange,
     time_ranges,
   });
-
-  // const { timeRangeRadioButtonsComponent, selectedTimeRangeButton } = TimeRangeRadioButtonsOnChange({
-  //   defaultTimeRange,
-  //   time_ranges,
-  // });
 
   const topArtistsFullList = useGetTopArtistsFullList(selectedTimeRangeButton);
 
@@ -48,7 +81,7 @@ export function UserTopArtistService({ pageSize, time_ranges, defaultTimeRange }
   );
 
   // artist list display logic
-  const { topArtistListOnChangeComponent } = TopArtistListOnChange({
+  const { topArtistListOnChangeComponent } = TopArtistListRecordsOnChange({
     currentArtists,
     startIndex,
   });
@@ -58,42 +91,4 @@ export function UserTopArtistService({ pageSize, time_ranges, defaultTimeRange }
     topArtistListOnChangeComponent,
     paginationComponent,
   };
-}
-////////////////////////////
-// Presentation Component //
-////////////////////////////
-interface UserTopArtistPresentationProps {
-  pageSize: number;
-  defaultTimeRange: string;
-  time_ranges: string[];
-}
-
-export function UserTopArtistPresentation({ pageSize, time_ranges, defaultTimeRange }: UserTopArtistPresentationProps) {
-  const { timeRangeRadioButtonsComponent, topArtistListOnChangeComponent, paginationComponent } = UserTopArtistService({
-    pageSize,
-    time_ranges,
-    defaultTimeRange,
-  });
-  const userTopArtist = (
-    <div className="top-artist-container">
-      <header className="top-artist-container-header">
-        Top Artists
-        {timeRangeRadioButtonsComponent}
-      </header>
-      <main className="top-artist-list-container">{topArtistListOnChangeComponent}</main>
-      <footer className="top-artist-container-footer">{paginationComponent}</footer>
-    </div>
-  );
-
-  return { userTopArtist };
-}
-
-interface UserTopArtistProps {
-  pageSize: number;
-  defaultTimeRange: string;
-  time_ranges: string[];
-}
-export function UserTopArtist({ pageSize, time_ranges, defaultTimeRange }: UserTopArtistProps) {
-  const { userTopArtist } = UserTopArtistPresentation({ pageSize, time_ranges, defaultTimeRange });
-  return {userTopArtist} ;
 }
